000 ARCHIVE/requirements.txt
python-amazon-sp-api>=0.17.0
pandas>=1.5.0
python-dotenv>=0.19.0